new-features

+++ home/views.py
@@ -13,7 +13,7 @@ def parking_places(request):
             return redirect('parking_places')
     else:
         form = ParkingPlaceForm()
-
+    
     parking_places = ParkingPlace.objects.all()
     
     return render(request, 'parking_place.html', {
@@ -21,14 +21,29 @@ def parking_places(request):
         "parking_places": parking_places
     })
 
+def edit_parking_place(request, pk):
+    parking_place = get_object_or_404(ParkingPlace, pk=pk)
+    if request.method == "POST":
+        form = ParkingPlaceForm(request.POST, instance=parking_place)
+        if form.is_valid():
+            form.save()
+            return redirect('parking_places')
+    else:
+        form = ParkingPlaceForm(instance=parking_place)
+    
+    return render(request, 'edit_parking_place.html', {'form': form})
+
 def delete_parking_place(request, pk):
     parking_place = get_object_or_404(ParkingPlace, pk=pk)
-    parking_place.delete()
-    return redirect('parking_places')
+    if request.method == "POST":
+        parking_place.delete()
+        return redirect('parking_places')
+    return render(request, 'delete_parking_place.html', {'parking_place': parking_place})
 
 def payments(request):
     payments = PaymentDetails.objects.all()
     return render(request, 'payments.html', {"payments": payments})
+
 
 def logs(request):
     logs = LogDetails.objects.all()
